Add build system for cyrenit and services
Add a GNU make based build system composed of a Makefile to build both
the binaries and an initcpio plus start a qemu instance to test the
system, an auxiliar script to ship depending libraries from the host
system.

Also, add a Makefile to the `services` directory where services for
testing purposes or internal usage will be built.

Signed-off-by: Ágatha Isabelle Chris Moreira Guedes <code@agatha.dev>